zeromq plugin: moved all the copy/pasted code in its own file
